Implement TensorRT-RTX inference
